TSDK-789 Reorg code
