+++ topl-btc-bridge/src/main/scala/co/topl/bridge/ToplBTCBridgeParamConfig.scala
@@ -15,6 +15,8 @@ case class ToplBTCBridgeParamConfig(
     toplWalletSeedFile: String = "topl-wallet.json",
     toplWalletPassword: String = "password",
     toplWalletDb: String = "topl-wallet.db",
+    btcHost: String = "localhost",
+    btcPort: Int = 18332,
     btcNetwork: BitcoinNetworkIdentifiers = RegTest,
     toplNetwork: ToplNetworkIdentifiers = ToplPrivatenet,
     toplHost: String = "localhost",
